% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{kpi_dcu}
\alias{kpi_dcu}
\title{KPI for daily count of users}
\usage{
kpi_dcu(con = intendo::sql_connection(), dates, customer = TRUE)
}
\arguments{
\item{con}{SQL connection}

\item{dates}{a vector (character) with YYYY-MM-DD}

\item{customer}{TRUE or FALSE (i.e. FALSE = free version of the game)}
}
\description{
Dplyr statements to obtain the metric
}
