% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{official_palettes}
\alias{official_palettes}
\title{Official Intendo Branches' Palettes}
\usage{
official_palettes(theme = "marketing")
}
\arguments{
\item{theme}{marketing or finance}
}
\value{
a vector with hex codes
}
\description{
Add description
}
